Commented range calculation out
I've commented out almost all of the code in the calculate_distances() function.
I wanted to test if I could remove some of the calculations.
I've now "tuned" the color filter to filter out the green of the lasers to an extent that will make sense.
Currently this part of the code will execute with about 160 Hz on my 2012 13" MacBook Pro.